requirements.txt
django>=4
django-ckeditor>=6.4.2
djangorestframework==3.12
djangorestframework-simplejwt>=5.2.0
django-cors-headers==3.7.0
django-filter>=22.1.0
django-environ>=0.9.0
gunicorn>=20.0.4,<20.1
markdown==3.1.1
psycopg2-binary>=2.8.5
PyJWT==1.7.1
Pillow>=7.1.2
pytz>=2019.3